Performance tweak in SecureRandomBase

+++ lib/random/secure_random_base.dart
@@ -34,7 +34,7 @@ abstract class SecureRandomBase implements SecureRandom {
   Uint8List nextBytes( int count ) {
     var bytes = new Uint8List(count);
     for( var i=0 ; i<count ; i++ ) {
-      bytes[i] = nextUint8().toInt();
+      bytes[i] = nextUint8();
     }
     return bytes;
   }
@@ -50,7 +50,7 @@ abstract class SecureRandomBase implements SecureRandom {
     // Generate random bytes and mask out any excess bits
     if (numBytes > 0) {
       for( var i=0 ; i<numBytes ; i++ ){
-        randomBits[i] = nextUint8().toInt();
+        randomBits[i] = nextUint8();
       }
       int excessBits = 8*numBytes - numBits;
       randomBits[0] &= (1 << (8-excessBits)) - 1;
